+++ NEWS.md
@@ -1,5 +1,7 @@
 # glycoverse (development version)
 
+* `glyfun` is now managed as a non-core glycoverse package.
+
 # glycoverse 0.3.0
 
 ## New features